# src/swagger.yaml
openapi: 3.0.0
info:
  title: AI Sales API
  description: Complete API documentation for the AI Sales system.
  version: 1.0.0
security:
  - BearerAuth: []
paths:
...
  /api/v1/interaction-modes:
    get:
      tags: [Interaction Modes]
      summary: Get all interaction modes
      description: Retrieve all interaction modes.
      responses:
        '200':
          description: Interaction modes retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InteractionModeResponse'
    post:
      tags: [Interaction Modes]
      summary: Create a new interaction mode
      description: Create a new interaction mode.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionModeCreate'
      responses:
        '201':
          description: Interaction mode created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionModeResponse'
  /api/v1/interaction-modes/{mode_id}:
    get:
      tags: [Interaction Modes]
      summary: Get interaction mode by ID
      description: Retrieve an interaction mode by its ID.
      parameters:
        - name: mode_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Interaction mode retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionModeResponse'
        '404':
          description: Interaction mode not found
...
  /api/v1/manufacturing-models/:
    post:
      tags: [Manufacturing Models]
      summary: Create a new manufacturing model
      description: Create a new manufacturing model in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufacturingModelCreate'
      responses:
        '201':
          description: Manufacturing model created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManufacturingModelResponse'
    get:
      tags: [Manufacturing Models]
      summary: Get all manufacturing models
      description: Retrieve all manufacturing models in the system.
      responses:
        '200':
          description: Manufacturing models retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ManufacturingModelResponse'
  /api/v1/manufacturing-models/{manufacturing_model_id}:
    get:
      tags: [Manufacturing Models]
      summary: Get manufacturing model by ID
      description: Retrieve a manufacturing model by its unique ID.
      parameters:
        - name: manufacturing_model_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Manufacturing model retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManufacturingModelResponse'
        '404':
          description: Manufacturing model not found
    put:
      tags: [Manufacturing Models]
      summary: Update manufacturing model
      description: Update a manufacturing model by its unique ID.
      parameters:
        - name: manufacturing_model_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManufacturingModelUpdate'
      responses:
        '200':
          description: Manufacturing model updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManufacturingModelResponse'
        '404':
          description: Manufacturing model not found
    delete:
      tags: [Manufacturing Models]
      summary: Delete manufacturing model
      description: Delete a manufacturing model by its unique ID.
      parameters:
        - name: manufacturing_model_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Manufacturing model deleted successfully
        '404':
          description: Manufacturing model not found

  # Plant Size Impacts Routes
  /api/v1/plant-size-impacts/:
    post:
      tags: [Plant Size Impacts]
      summary: Create a new plant size impact
      description: Create a new plant size impact in the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantSizeImpactCreate'
      responses:
        '201':
          description: Plant size impact created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantSizeImpactResponse'
    get:
      tags: [Plant Size Impacts]
      summary: Get all plant size impacts
      description: Retrieve all plant size impacts in the system.
      responses:
        '200':
          description: Plant size impacts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlantSizeImpactResponse'
  /api/v1/plant-size-impacts/{plant_size_impact_id}:
    get:
      tags: [Plant Size Impacts]
      summary: Get plant size impact by ID
      description: Retrieve a plant size impact by its unique ID.
      parameters:
        - name: plant_size_impact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plant size impact retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantSizeImpactResponse'
        '404':
          description: Plant size impact not found
    put:
      tags: [Plant Size Impacts]
      summary: Update plant size impact
      description: Update a plant size impact by its unique ID.
      parameters:
        - name: plant_size_impact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlantSizeImpactUpdate'
      responses:
        '200':
          description: Plant size impact updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantSizeImpactResponse'
        '404':
          description: Plant size impact not found
    delete:
      tags: [Plant Size Impacts]
      summary: Delete plant size impact
      description: Delete a plant size impact by its unique ID.
      parameters:
        - name: plant_size_impact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Plant size impact deleted successfully
        '404':
          description: Plant size impact not found
...
  /api/v1/interaction-mode-plant-size-impacts:
    post:
      tags: [Interaction Mode Plant Size Impacts]
      summary: Create an Interaction Mode Plant Size Impact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionModePlantSizeImpactBase'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionModePlantSizeImpactResponse'
    get:
      tags: [Interaction Mode Plant Size Impacts]
      summary: Get all Interaction Mode Plant Size Impacts
      responses:
        '200':
          description: List of Interaction Mode Plant Size Impacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InteractionModePlantSizeImpactResponse'
  /api/v1/interaction-mode-plant-size-impacts/{id}:
    get:
      tags: [Interaction Mode Plant Size Impacts]
      summary: Get Interaction Mode Plant Size Impact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Interaction Mode Plant Size Impact
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionModePlantSizeImpactResponse'
        '404':
          description: Not found
    put:
      tags: [Interaction Mode Plant Size Impacts]
      summary: Update Interaction Mode Plant Size Impact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionModePlantSizeImpactBase'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionModePlantSizeImpactResponse'
        '404':
          description: Not found
    delete:
      tags: [Interaction Mode Plant Size Impacts]
      summary: Delete Interaction Mode Plant Size Impact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InteractionModeCreate:
      type: object
      required:
        - name
        - description
        - prompt_template
      properties:
        name:
          type: string
          enum: [prospecting, sales, closing]
        description:
          type: string
        prompt_template:
          type: string
    InteractionModeResponse:
      type: object
      properties:
        mode_id:
          type: string
          format: uuid
        name:
          type: string
          enum: [prospecting, sales, closing]
        description:
          type: string
        prompt_template:
          type: string
        status_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ManufacturingModelBase:
      type: object
      properties:
        name:
          type: string
    ManufacturingModelCreate:
      allOf:
        - $ref: '#/components/schemas/ManufacturingModelBase'
    ManufacturingModelUpdate:
      type: object
      properties:
        name:
          type: string
          nullable: true
    ManufacturingModelResponse:
      allOf:
        - $ref: '#/components/schemas/ManufacturingModelBase'
        - type: object
          properties:
            manufacturing_model_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    PlantSizeImpactBase:
      type: object
      properties:
        name:
          type: string
    PlantSizeImpactCreate:
      allOf:
        - $ref: '#/components/schemas/PlantSizeImpactBase'
    PlantSizeImpactUpdate:
      type: object
      properties:
        name:
          type: string
          nullable: true
    PlantSizeImpactResponse:
      allOf:
        - $ref: '#/components/schemas/PlantSizeImpactBase'
        - type: object
          properties:
            plant_size_impact_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...
    InteractionModePlantSizeImpactBase:
      type: object
      properties:
        mode_id:
          type: string
        plant_size_impact_id:
          type: string
        prompt_template:
          type: string
    InteractionModePlantSizeImpactResponse:
      allOf:
        - $ref: '#/components/schemas/InteractionModePlantSizeImpactBase'
        - type: object
          properties:
            interaction_mode_plant_size_impact_id:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...